it's really works wow

+++ ai-interview-service/src/main/java/io/github/kstnnn/ai/interview/service/service/TopicStateService.java
@@ -9,7 +9,7 @@ import java.util.UUID;
 public interface TopicStateService {
 
   SessionTopicState updateAfterAnswer(
-      UUID sessionId, String topic, int totalScore, double aiConfidence, int roundNumber);
+      UUID sessionId, String topic, double totalScore, double aiConfidence, int roundNumber);
 
   double calculateSessionConfidence(UUID sessionId);
